chore: out channel fix
