+++ pkg/ethereum/node.go
@@ -3,8 +3,6 @@ package ethereum
 import (
 	"context"
 	"fmt"
-	"log/slog"
-	"net"
 	"os"
 	"sync"
 	"time"
@@ -18,7 +16,6 @@ import (
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/libp2p/go-libp2p/p2p/security/noise"
 	"github.com/libp2p/go-libp2p/p2p/transport/tcp"
-	ma "github.com/multiformats/go-multiaddr"
 	"github.com/nats-io/nats.go"
 	"github.com/nats-io/nats.go/jetstream"
 	"github.com/pkg/errors"
@@ -275,26 +272,4 @@ func (n *Node) getMetadataFromCache(pid peer.ID) (*PeerMetadata, error) {
 	}
 
 	return metadata, nil
-}
-
-// MaddrFrom takes in an ip address string and port to produce a go multiaddr format.
-func MaddrFrom(ip string, port uint) (ma.Multiaddr, error) {
-	parsed := net.ParseIP(ip)
-	if parsed == nil {
-		return nil, fmt.Errorf("invalid IP address: %s", ip)
-	} else if parsed.To4() != nil {
-		return ma.NewMultiaddr(fmt.Sprintf("/ip4/%s/tcp/%d", ip, port))
-	} else if parsed.To16() != nil {
-		return ma.NewMultiaddr(fmt.Sprintf("/ip6/%s/tcp/%d", ip, port))
-	} else {
-		return nil, fmt.Errorf("invalid IP address: %s", ip)
-	}
-}
-
-func LogAttrPeerID(pid peer.ID) slog.Attr {
-	return slog.String("AttrKeyPeerID", pid.String())
-}
-
-func LogAttrError(err error) slog.Attr {
-	return slog.Attr{Key: "AttrKeyError", Value: slog.AnyValue(err)}
 }
